local_cache: Don't start multiple housekeeping processes

+++ src/cache/local_cache.cpp
@@ -69,6 +69,7 @@ const std::string DIRECT_CACHE_MANIFEST_FILE_NAME = ".manifest";
 const std::string CACHE_ENTRY_FILE_NAME = ".entry";
 const std::string FILE_LOCK_SUFFIX = ".lock";
 const std::string STATS_FILE_NAME = "stats.json";
+const std::string HOUSEKEEPING_FILE_LOCK = ".housekeeping" + FILE_LOCK_SUFFIX;
 
 // Maximum number of manifests per direct mode cache entry (must be at least 1). Set this too low,
 // and there will be cache thrashing (e.g. when switching branches). Set this too high and cache
@@ -205,7 +206,7 @@ void purge_old_cache_entries(const std::string& root_folder) {
         // We acquire a scoped lock for the cache entry before deleting it.
         const auto file_lock_path = cache_entry_file_lock_path(dir.path());
         {
-          file::file_lock_t lock(file_lock_path, config::remote_locks());
+          file_lock_t lock{file_lock_path, file_lock_t::to_remote_t(config::remote_locks())};
           if (lock.has_lock()) {
             file::remove_dir(dir.path());
             total_size -= dir.size();
@@ -249,7 +250,7 @@ void delete_stale_lock_files(const std::string& root_folder) {
       for (const auto& info : files) {
         if (is_potentially_stale_lock_file(info, now)) {
           // We consider the lock to be stale if we can get the lock.
-          file::file_lock_t lock(info.path(), false);
+          file_lock_t lock{info.path(), file_lock_t::remote_t::YES};
           if (lock.has_lock()) {
             // Delete the file if it is stale.
             debug::log(debug::DEBUG) << "Deleting stale " << info.path();
@@ -306,7 +307,7 @@ void local_cache_t::clear() {
       // We acquire an exclusive lock for the cache entry before deleting it.
       const auto file_lock_path = cache_entry_file_lock_path(dir.path());
       {
-        file::file_lock_t lock(file_lock_path, config::remote_locks());
+        file_lock_t lock{file_lock_path, file_lock_t::to_remote_t(config::remote_locks())};
         if (lock.has_lock()) {
           file::remove_dir(dir.path());
         }
@@ -328,19 +329,28 @@ void local_cache_t::clear() {
 }
 
 void local_cache_t::perform_housekeeping() {
-  const auto start_t = std::chrono::high_resolution_clock::now();
+  // Use a file lock to prevent multiple concurrent housekeeping processes.
+  const auto file_lock_path = file::append_path(config::dir(), HOUSEKEEPING_FILE_LOCK);
+  file_lock_t lock{file_lock_path,
+                   file_lock_t::to_remote_t(config::remote_locks()),
+                   file_lock_t::blocking_t::NO};
+  if (lock.has_lock()) {
+    debug::log(debug::INFO) << "Performing housekeeping.";
 
-  debug::log(debug::INFO) << "Performing housekeeping.";
+    const auto start_t = std::chrono::high_resolution_clock::now();
 
-  // Purge old cache entries.
-  purge_old_cache_entries(config::dir());
+    // Purge old cache entries.
+    purge_old_cache_entries(config::dir());
 
-  // Delete old stale lock files.
-  delete_stale_lock_files(config::dir());
+    // Delete old stale lock files.
+    delete_stale_lock_files(config::dir());
 
-  const auto stop_t = std::chrono::high_resolution_clock::now();
-  const auto dt = std::chrono::duration_cast<std::chrono::milliseconds>(stop_t - start_t).count();
-  debug::log(debug::INFO) << "Finished housekeeping in " << dt << " ms";
+    const auto stop_t = std::chrono::high_resolution_clock::now();
+    const auto dt = std::chrono::duration_cast<std::chrono::milliseconds>(stop_t - start_t).count();
+    debug::log(debug::INFO) << "Finished housekeeping in " << dt << " ms";
+  } else {
+    debug::log(debug::INFO) << "Skipping housekeeping (it is already in progress).";
+  }
 }
 
 void local_cache_t::show_stats() {
@@ -359,7 +369,8 @@ void local_cache_t::show_stats() {
     visited_dirs.insert(first_level_dir_path);
     const auto stats_path = file::append_path(first_level_dir_path, STATS_FILE_NAME);
     cache_stats_t stats;
-    file::file_lock_t lock{stats_path + FILE_LOCK_SUFFIX, config::remote_locks()};
+    file_lock_t lock{stats_path + FILE_LOCK_SUFFIX,
+                     file_lock_t::to_remote_t(config::remote_locks())};
     if (!lock.has_lock()) {
       debug::log(debug::DEBUG) << "Failed to lock stats, skipping";
       return;
@@ -398,7 +409,8 @@ void local_cache_t::zero_stats() {
   for (const auto& dir : dirs) {
     try {
       const auto stats_path = file::append_path(dir.path(), STATS_FILE_NAME);
-      file::file_lock_t lock{stats_path + FILE_LOCK_SUFFIX, config::remote_locks()};
+      file_lock_t lock{stats_path + FILE_LOCK_SUFFIX,
+                       file_lock_t::to_remote_t(config::remote_locks())};
       if (lock.has_lock()) {
         file::remove_file(stats_path);
       }
@@ -492,7 +504,8 @@ void local_cache_t::add(const std::string& hash,
 
   {
     // Acquire a scoped exclusive lock for the cache entry.
-    file::file_lock_t lock(cache_entry_file_lock_path(cache_entry_path), config::remote_locks());
+    file_lock_t lock{cache_entry_file_lock_path(cache_entry_path),
+                     file_lock_t::to_remote_t(config::remote_locks())};
     if (!lock.has_lock()) {
       throw std::runtime_error("Unable to acquire a cache entry lock for writing.");
     }
@@ -527,7 +540,7 @@ void local_cache_t::add(const std::string& hash,
   }
 }
 
-std::pair<cache_entry_t, file::file_lock_t> local_cache_t::lookup(const std::string& hash) {
+std::pair<cache_entry_t, file_lock_t> local_cache_t::lookup(const std::string& hash) {
   // Get the path to the cache entry.
   const auto cache_entry_path = hash_to_cache_entry_path(hash);
 
@@ -541,7 +554,8 @@ std::pair<cache_entry_t, file::file_lock_t> local_cache_t::lookup(const std::str
     }
 
     // Acquire a scoped lock for the cache entry.
-    file::file_lock_t lock(cache_entry_file_lock_path(cache_entry_path), config::remote_locks());
+    file_lock_t lock{cache_entry_file_lock_path(cache_entry_path),
+                     file_lock_t::to_remote_t(config::remote_locks())};
     if (!lock.has_lock()) {
       throw std::runtime_error("Unable to acquire a cache entry lock for reading.");
     }
@@ -554,7 +568,7 @@ std::pair<cache_entry_t, file::file_lock_t> local_cache_t::lookup(const std::str
     return std::make_pair(cache_entry_t::deserialize(entry_data), std::move(lock));
   } catch (...) {
     update_stats(hash, cache_stats_t::local_miss());
-    return std::make_pair(cache_entry_t(), file::file_lock_t());
+    return std::make_pair(cache_entry_t(), file_lock_t());
   }
 }
 
@@ -568,7 +582,8 @@ bool local_cache_t::update_stats(const std::string& hash,
       file::create_dir_with_parents(cache_subdir);
     }
     const auto stats_file_path = file::append_path(cache_subdir, STATS_FILE_NAME);
-    file::file_lock_t lock(stats_file_path + FILE_LOCK_SUFFIX, config::remote_locks());
+    file_lock_t lock{stats_file_path + FILE_LOCK_SUFFIX,
+                     file_lock_t::to_remote_t(config::remote_locks())};
     if (!lock.has_lock()) {
       debug::log(debug::INFO) << "Failed to lock stats, skipping update";
       return false;
